price-polling/requirements.txt
yfinance
flask
requests
pandas
mysql-connector-python
